requirements.txt
transformers 
sentence-transformers 
faiss-cpu 
faster-whisper 
TTS 
flask 
flask-socketio